+++ CHANGELOG.md
@@ -3,11 +3,12 @@
 ## TODO LIST
 - SignAuthInterceptor  使用SDK  的sample和例子，简化对接成本
 - Elasticsearch 深入研究，面试，实战
-- 代码在线生成器
+- [代码在线生成器](https://gitee.com/zhangquansheng/magic/tree/4.2.0-springboot-code-gen/)
 - SOLID 架构设计 深入学习
 - cache aside pattern / redis 深入总结
 - 技术分享整理
-- 增加zk分布式锁
+- 增加zk分布式锁，分布式锁实现方式的性能对比和使用场景整理
+- 如何保证kafka消息不丢失（结合实际业务场景）
 
 ### 强烈提醒
 - **需要花费很长时间的事情，需要慎重考虑一下在去做**